+++ app/src/test/java/com/mysugr/android/testing/v1/example/auth/AuthManagerMockSteps.kt
@@ -5,7 +5,7 @@ import com.mysugr.android.testing.example.auth.AuthManager.LoginOrRegisterResult
 import com.mysugr.android.testing.util.nonNullableAny
 import com.mysugr.sweetest.framework.base.BaseSteps
 import com.mysugr.sweetest.framework.base.dependency
-import com.mysugr.sweetest.TestContext
+import dev.sweetest.internal.TestContext
 import org.mockito.Mockito.`when`
 import org.mockito.Mockito.anyString
 import org.mockito.Mockito.mock
