+++ sweetest/api/v1/src/test/java/com/mysugr/sweetest/DependenciesTest.kt
@@ -8,6 +8,7 @@ import com.mysugr.sweetest.framework.base.steps
 import com.mysugr.sweetest.framework.configuration.ModuleTestingConfiguration
 import com.mysugr.sweetest.framework.configuration.moduleTestingConfiguration
 import com.mysugr.sweetest.util.isMock
+import dev.sweetest.internal.TestContext
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
